Revert "Merge branch 'doc_update_core_bigint' of github.com:CodemateLtd/sdk into doc_update_core_bigint"
This reverts commit 95b70be0b4146a786399ffb33613ac4de9ca6959, reversing
changes made to 6cde67faec232661f9bf2954cf73b353f1ad7e7d.

+++ pkg/dartdev/test/utils.dart
@@ -93,21 +93,11 @@ dev_dependencies:
     file.deleteSync();
   }
 
-  Future<void> dispose() async {
+  void dispose() {
     _process?.kill();
-    await _process?.exitCode;
     _process = null;
-    int deleteAttempts = 5;
-    while (dir.existsSync()) {
-      try {
-        dir.deleteSync(recursive: true);
-      } catch (e) {
-        if ((--deleteAttempts) <= 0) {
-          rethrow;
-        }
-        await Future.delayed(Duration(milliseconds: 500));
-        print('Got $e while deleting $dir. Trying again...');
-      }
+    if (dir.existsSync()) {
+      dir.deleteSync(recursive: true);
     }
   }
 
@@ -145,8 +135,7 @@ dev_dependencies:
           ...arguments,
         ],
         workingDirectory: workingDir ?? dir.path,
-        environment: {if (logAnalytics) '_DARTDEV_LOG_ANALYTICS': 'true'})
-      ..then((p) => _process = p);
+        environment: {if (logAnalytics) '_DARTDEV_LOG_ANALYTICS': 'true'});
   }
 
   String _sdkRootPath;
